Refactored out Ising model specific functions to it's own C++ and header files, and moved the Problem 4 specific code of the main method to it's own file.
