+++ Project-4/Problem-4/README.md
@@ -2,12 +2,12 @@
 Implement Ising model with Monte Carlo.
 
 ## Usage
-- Compile the code with `g++ ising.cpp -o ising.exe ../src/* -I../include/ -std=c++11 -larmadillo`
-- Run the code with `./ising.exe <L width of lattice> <T temperature in J/Kb> <Number MC cycles>`
+- Compile the code with `g++ problem4.cpp -o problem4.exe ../src/* -I../include/ -std=c++11 -larmadillo`
+- Run the code with `./problem4.exe <L width of lattice> <T temperature in J/Kb> <Number MC cycles>`
 
 Examples on command lines 
-`./ising.exe 2`
-`./ising.exe 2 1 100`
-`./ising.exe 50 1 1000`
-`./ising.exe 50 1 10000`
+`./problem4.exe 2`
+`./problem4.exe 2 1 100`
+`./problem4.exe 50 1 1000`
+`./problem4.exe 50 1 10000`
